+++ src/r_things.c
@@ -741,8 +741,15 @@ static void R_ProjectSprite(mobj_t *thing)
     else
         vis->shadowpos = 1;
 
-    if ((thing->flags & MF_FUZZ) && pausesprites)
-        vis->colfunc = (r_textures ? R_DrawPausedFuzzColumn : thing->colfunc);
+    if (thing->flags & MF_FUZZ)
+    {
+        if (r_blood == r_blood_colors && thing->type == MT_FUZZYBLOOD)
+            vis->colfunc = (r_translucency ? R_DrawTranslucent33Column : R_DrawColumn);
+        else if (pausesprites)
+            vis->colfunc = (r_textures ? R_DrawPausedFuzzColumn : thing->colfunc);
+        else
+            vis->colfunc = (invulnerable && r_textures ? thing->altcolfunc : thing->colfunc);
+    }
     else
         vis->colfunc = (invulnerable && r_textures ? thing->altcolfunc : thing->colfunc);
 
@@ -866,6 +873,11 @@ static void R_ProjectBloodSplat(const bloodsplat_t *splat)
     else if (r_blood == r_blood_red)
     {
         vis->blood = MT_BLOOD;
+        vis->colfunc = (r_bloodsplats_translucency ? R_DrawBloodSplatColumn : R_DrawSolidBloodSplatColumn);
+    }
+    else if (r_blood == r_blood_colors)
+    {
+        vis->blood = (splat->colfunc == fuzzcolfunc ? MT_BLOOD : splat->blood);
         vis->colfunc = (r_bloodsplats_translucency ? R_DrawBloodSplatColumn : R_DrawSolidBloodSplatColumn);
     }
     else